Changes so far
train.py - adjusted so that modifying network to work with different sampling rates will work correctly (so long as sample_rate*sample_length = 144000)

analyze_MZ.py- a modified version of analyze.py that includes the ability to set --overwrite. This skips files in the output directory that have already been analyzed (if set to 0). This is useful if you're running birdnet on a very large file set and often have to stop or pause running code. 

embeddings_MZ.py- a modified version of embeddings.py that includes --overwrite as well as the ability to get embeddings from a custom classifier. The base version on BirdNET-Analyzer does not currently have this capability. 

config.py- modified for use with the other files included here.